requirements/base.txt
django
djangorestframework
mysqlclient
psycopg2
boto3
django-storages
